Refactor FST loader into GwLoader subclass
