+++ src/dump_file_main.h
@@ -1,0 +1,4 @@
+#pragma once
+
+GwDumpFile *vcd_recoder_main(char *fname);
+GwDumpFile *fst_main(char *fname, char *skip_start, char *skip_end);
