+++ src/gw-fst-util.h
