+++ src/lx2.c
@@ -22,7 +22,6 @@
 #include "debug.h"
 #include "busy.h"
 #include "hierpack.h"
-#include "fst.h"
 #include "gw-vcd-file.h"
 #include "gw-fst-file.h"
