+++ src/meson.build
@@ -38,17 +38,17 @@ gtkwave_sources = [
     'debug.c',
     'edgebuttons.c',
     'entry.c',
+    'dump_file_main.c',
     'fetchbuttons.c',
     'fgetdynamic.c',
     'file.c',
     'fonts.c',
-    'fst_util.c',
-    'fst.c',
     'gconf.c',
     'ghw.c',
     'globals.c',
     'gtk23compat.c',
     'gw-fst-file.c',
+    'gw-fst-loader.c',
     'gw-ghw-file.c',
     'gw-mouseover.c',
     'gw-named-marker-dialog.c',
