+++ src/vcd.c
@@ -52,20 +52,3 @@ void strcpy_vcdalt(char *too, char *from, char delim)
         }
     } while ((*(too++) = ch));
 }
-
-// TODO: remove
-GwDumpFile *vcd_recoder_main(char *fname)
-{
-    const Settings *global_settings = &GLOBALS->settings;
-
-    GwLoader *loader = gw_vcd_loader_new();
-    gw_loader_set_preserve_glitches(loader, global_settings->preserve_glitches);
-    gw_loader_set_preserve_glitches_real(loader, global_settings->preserve_glitches_real);
-    gw_vcd_loader_set_vlist_prepack(GW_VCD_LOADER(loader), global_settings->vlist_prepack);
-
-    GwDumpFile *file = gw_loader_load(loader, fname, NULL); // TODO: use error
-
-    g_object_unref(loader);
-
-    return file;
-}
